Update README instructions
`pyvenv` was deprecated in Python 3.6:
https://docs.python.org/dev/whatsnew/3.6.html#id8

Something has I think changed with pathing such that you can no longer
just import the package by running `zoom` from the base of the project.
Rather than figure that out and potentially restructuring the app, I
just added the commands to run the install in "editable" mode, which is
an extra step but works just fine.